requirements.txt
flask
line-bot-sdk==3.*
openai
supabase-py
python-dotenv
